requirements.txt
fastapi
pydantic
motor
uvicorn
python-dotenv
passlib[bcrypt]
pyjwt